Cli logic moved to CliProcessorBase.
packages consolidated and updated.
WorkflowLoggerHelper splitted on WorkflowLoggerHelper and WorkflowLoggerHelperBase.
CommonUtils: added ParseEnum.
