+++ Sources/PT.PM.AntlrUtils/AntlrError.cs
@@ -1,9 +1,4 @@
 ﻿using Antlr4.Runtime;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace PT.PM.AntlrUtils
 {
