+++ Sources/PT.PM.CSharpParseTreeUst/packages.config
@@ -1,8 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Microsoft.CodeAnalysis.Analyzers" version="1.1.0" targetFramework="net45" />
-  <package id="Microsoft.CodeAnalysis.Common" version="2.2.0" targetFramework="net462" />
-  <package id="Microsoft.CodeAnalysis.CSharp" version="2.2.0" targetFramework="net462" />
+  <package id="Microsoft.CodeAnalysis.Common" version="2.4.0" targetFramework="net462" />
+  <package id="Microsoft.CodeAnalysis.CSharp" version="2.4.0" targetFramework="net462" />
   <package id="System.AppContext" version="4.3.0" targetFramework="net462" />
   <package id="System.Collections" version="4.3.0" targetFramework="net462" />
   <package id="System.Collections.Concurrent" version="4.3.0" targetFramework="net462" />
