+++ Sources/PT.PM.Cli/CliProcessor.cs
@@ -1,0 +1,47 @@
+﻿using System.Collections.Generic;
+using System.Diagnostics;
+using PT.PM.Common;
+using PT.PM.Common.CodeRepository;
+using PT.PM.Matching;
+using PT.PM.Matching.PatternsRepository;
+
+namespace PT.PM.Cli
+{
+    public class CliProcessor : CliProcessorBase<Stage, WorkflowResult, PatternRoot, MatchingResult>
+    {
+        protected override WorkflowResult InitWorkflowAndProcess(CliParameters parameters, ILogger logger, SourceCodeRepository sourceCodeRepository, IPatternsRepository patternsRepository, out Stopwatch stopwatch)
+        {
+            Stage stage = string.IsNullOrEmpty(parameters.InputFileNameOrDirectory)
+                ? Stage.Pattern
+                : parameters.Stage.ParseEnum(Stage.Match);
+
+            var workflow = new Workflow(sourceCodeRepository, patternsRepository, stage)
+            {
+                Logger = logger,
+                ThreadCount = parameters.ThreadCount,
+                MemoryConsumptionMb = parameters.Memory,
+                IsIncludePreprocessing = parameters.IsPreprocessUst,
+                LogsDir = parameters.LogsDir,
+                DumpDir = parameters.LogsDir,
+                StartStage = parameters.StartStage.ParseEnum(Stage.File),
+                DumpStages = new HashSet<Stage>(parameters.DumpStages.ParseCollection<Stage>()),
+                IndentedDump = parameters.IndentedDump,
+                DumpWithTextSpans = parameters.IncludeTextSpansInDump,
+                RenderStages = new HashSet<Stage>(parameters.RenderStages.ParseCollection<Stage>()),
+                RenderFormat = parameters.RenderFormat.ParseEnum<GraphvizOutputFormat>(),
+                RenderDirection = parameters.RenderDirection.ParseEnum<GraphvizDirection>(),
+            };
+            stopwatch = Stopwatch.StartNew();
+            WorkflowResult workflowResult = workflow.Process();
+            stopwatch.Stop();
+
+            return workflowResult;
+        }
+
+        protected override void LogStatistics(ILogger logger, WorkflowResult workflowResult)
+        {
+            var workflowLoggerHelper = new WorkflowLoggerHelper(logger, workflowResult);
+            workflowLoggerHelper.LogStatistics();
+        }
+    }
+}
