+++ Sources/PT.PM.Cli/CliProcessorBase.cs
@@ -1,0 +1,151 @@
+﻿using CommandLine;
+using PT.PM.Common;
+using PT.PM.Common.CodeRepository;
+using PT.PM.Matching;
+using PT.PM.Matching.PatternsRepository;
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Globalization;
+using System.Linq;
+using System.Threading;
+
+namespace PT.PM.Cli
+{
+    public abstract class CliProcessorBase<TStage, TWorkflowResult, TPattern, TMatchingResult>
+        where TStage : struct, IConvertible
+        where TWorkflowResult : WorkflowResultBase<TStage, TPattern, TMatchingResult>
+        where TMatchingResult : MatchingResultBase<TPattern>
+    {
+        public int ParseAndConvert(string[] args)
+        {
+            Thread.CurrentThread.CurrentCulture = CultureInfo.InvariantCulture;
+            Thread.CurrentThread.CurrentUICulture = CultureInfo.InvariantCulture;
+
+            var parser = new Parser(config => config.HelpWriter = Console.Out);
+            ParserResult<CliParameters> parserResult = parser.ParseArguments<CliParameters>(args);
+
+            var result = parserResult.MapResult(
+                cliParams => {
+                    int convertResult = 0;
+                    if (cliParams.MaxStackSize == 0)
+                    {
+                        convertResult = Convert(args, cliParams);
+                    }
+                    else
+                    {
+                        Thread thread = new Thread(() =>
+                        {
+                            convertResult = Convert(args, cliParams);
+                        },
+                        cliParams.MaxStackSize);
+                        thread.Start();
+                        thread.Join();
+                    }
+                    return convertResult;
+                },
+                errors => ProcessErrors(errors));
+
+            return result;
+        }
+
+        public int Convert(string[] args, CliParameters parameters)
+        {
+            ILogger logger = new ConsoleFileLogger();
+
+            try
+            {
+                if (parameters.ShowVersion)
+                {
+                    string version = System.Reflection.Assembly.GetExecutingAssembly().GetName().Version.ToString();
+                    logger.LogInfo($"PT.PM version: {version}");
+                }
+
+                if (logger is FileLogger abstractLogger)
+                {
+                    string commandLineArguments = "Command line arguments" + (args.Length > 0
+                       ? ": " + string.Join(" ", args)
+                       : " are not defined.");
+                    abstractLogger.LogsDir = parameters.LogsDir;
+                    abstractLogger.IsLogErrors = parameters.IsLogErrors;
+                    abstractLogger.IsLogDebugs = parameters.IsLogDebugs;
+                    abstractLogger.LogInfo(commandLineArguments);
+                }
+
+                if (string.IsNullOrEmpty(parameters.InputFileNameOrDirectory) && string.IsNullOrEmpty(parameters.Patterns))
+                {
+                    throw new ArgumentException("at least --files or --patterns parameter required");
+                }
+
+                if (!Enum.TryParse(parameters.Stage, out Stage pmStage))
+                {
+                    pmStage = Stage.Match;
+                }
+                if (!Enum.TryParse(parameters.StartStage, out Stage pmStartStage))
+                {
+                    pmStartStage = Stage.File;
+                }
+
+                HashSet<Language> languages = parameters.Languages.ParseLanguages();
+                SourceCodeRepository sourceCodeRepository = RepositoryFactory.
+                    CreateSourceCodeRepository(parameters.InputFileNameOrDirectory, languages, parameters.TempDir,
+                    pmStartStage == Stage.Ust);
+
+                logger.SourceCodeRepository = sourceCodeRepository;
+
+                IPatternsRepository patternsRepository = RepositoryFactory.CreatePatternsRepository(parameters.Patterns);
+
+                Stopwatch stopwatch;
+                TWorkflowResult workflowResult =
+                    InitWorkflowAndProcess(parameters, logger, sourceCodeRepository, patternsRepository, out stopwatch);
+
+                if (pmStage != Stage.Pattern)
+                {
+                    logger.LogInfo("Scan completed.");
+                    if (pmStage == Stage.Match)
+                    {
+                        logger.LogInfo($"{"Matches count: ",-22} {workflowResult.MatchingResults.Count()}");
+                    }
+                }
+                else
+                {
+                    logger.LogInfo("Patterns checked.");
+                }
+                logger.LogInfo($"{"Errors count: ",-22} {workflowResult.ErrorCount}");
+                LogStatistics(logger, workflowResult);
+                logger.LogInfo($"{"Time elapsed:",-22} {stopwatch.Elapsed}");
+            }
+            catch (Exception ex)
+            {
+                if (logger != null)
+                {
+                    if (logger is FileLogger abstractLogger)
+                    {
+                        abstractLogger.IsLogErrors = true;
+                    }
+                    logger.LogError(ex);
+                }
+
+                return 1;
+            }
+            finally
+            {
+                if (logger is IDisposable disposableLogger)
+                {
+                    disposableLogger.Dispose();
+                }
+            }
+
+            return 0;
+        }
+
+        protected abstract TWorkflowResult InitWorkflowAndProcess(CliParameters parameters, ILogger logger, SourceCodeRepository sourceCodeRepository, IPatternsRepository patternsRepository, out Stopwatch stopwatch);
+
+        protected abstract void LogStatistics(ILogger logger, TWorkflowResult workflowResult);
+
+        protected int ProcessErrors(IEnumerable<Error> errors)
+        {
+            return 1;
+        }
+    }
+}
