+++ Sources/PT.PM.Cli/packages.config
@@ -7,12 +7,12 @@
   <package id="System.Console" version="4.3.0" targetFramework="net462" />
   <package id="System.Diagnostics.Debug" version="4.3.0" targetFramework="net462" />
   <package id="System.Globalization" version="4.3.0" targetFramework="net462" />
-  <package id="System.IO" version="4.1.0-rc2-24027" targetFramework="net462" />
+  <package id="System.IO" version="4.3.0" targetFramework="net462" />
   <package id="System.Linq" version="4.3.0" targetFramework="net462" />
   <package id="System.Linq.Expressions" version="4.3.0" targetFramework="net462" />
   <package id="System.Reflection" version="4.3.0" targetFramework="net462" />
-  <package id="System.Reflection.Extensions" version="4.0.1-rc2-24027" targetFramework="net462" />
-  <package id="System.Reflection.TypeExtensions" version="4.1.0-rc2-24027" targetFramework="net462" />
+  <package id="System.Reflection.Extensions" version="4.3.0" targetFramework="net462" />
+  <package id="System.Reflection.TypeExtensions" version="4.4.0" targetFramework="net462" />
   <package id="System.Resources.ResourceManager" version="4.3.0" targetFramework="net462" />
   <package id="System.Runtime" version="4.3.0" targetFramework="net462" />
   <package id="System.Runtime.Extensions" version="4.3.0" targetFramework="net462" />
